RPM and Puppet plugins now use %{version} for Requires
Previously, RPM and Puppet both had a special macro definition for the
platform version required. However, because both these plugins are
versioned in lockstep with the platform, we can just use the %{version}
instead.

Without this change, the Requires for each package would not be set
correctly by the rel-eng script, as it simply sets the Version field.